+++ 3D Project/TerrainScene.h
@@ -33,7 +33,7 @@ class TerrainScene : public Scene {
 		// Vertex buffer.
 		GLuint vertexBuffer = 0;
 		GLuint vertexAttribute = 0;
-		unsigned short vertexCount = 0;
+		unsigned int vertexCount = 0;
 
 		// Index buffer.
 		GLuint indexBuffer = 0;
